IG-20008: Improve performance of info()
Improve the performance of info() by using GetItem instead of listing the file's parent directory, and by only minimally listing a directory to check for its existence.

The performance of the original solution is O(N^2), while the new solution is O(N). I measured a 77% reduction in runtime when reading a directory containing 232 small parquet files.